requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.4
brotlipy==0.7.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama==0.4.6
crcmod==1.7
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
Cython==3.0.11
fonttools==4.38.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.2
importlib-metadata==6.7.0
jmespath==0.10.0
kiwisolver==1.4.5
Markdown==3.4.4
markdown-it-py==2.2.0
matplotlib==3.5.3
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
mmcv==1.7.0
mmcv-full==1.7.2
# Editable install with no version control (mmdet==2.25.2)
-e /home/estar/Xia/mmdetection-2.25.2
model-index==0.1.11
networkx==2.6.3
numpy==1.21.5
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.10
ordered-set==4.1.0
oss2==2.17.0
packaging==24.0
pandas==1.3.5
Pillow==9.4.0
platformdirs==4.0.0
pycocotools==2.0.7
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.20.0
Pygments==2.17.2
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.1.2
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.9.0.post0
pytz==2023.4
PyWavelets==1.3.0
PyYAML==6.0.1
requests==2.28.2
rich==13.7.1
scikit-image==0.19.3
scipy==1.7.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
tabulate==0.9.0
terminaltables==3.1.10
tifffile==2021.11.2
tomli==2.0.1
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.65.2
typing_extensions==4.7.1
urllib3 @ file:///croot/urllib3_1673575502006/work
yapf==0.40.1
zipp==3.15.0
